default users can't be deleted

+++ app/views/admin/admins.blade.php
@@ -123,9 +123,13 @@
 	            		<td>{{{ $admin->firstname }}}</td>
 	            		<td>{{{ $admin->middlename }}}</td>
 	            		<td><a href="{{URL::route('aadminstasks', $admin->id) }}" class="pull-right btn btn-default">More</a>
+                  @if($admin->employee_id != '1234567' || $admin->employee_id != '123456789')
 	            		{{ Form::open(['method'=> 'DELETE','route'=>['deleteadmin', $admin->id]]) }}
 	            		{{ Form::submit('Delete', ['class' => 'pull-right btn btn-danger','onclick'=>'return confirm("Continue to delete?")']) }}
 	            		{{ Form::close() }}
+                  @else
+                  <button type="button" class="btn btn-danger" onclick="return alert('Fuck You! Don\'t delete this.)">Delete</button>
+                  @endif
 	            		</td>
 	            	</tr>
 	                @endforeach
